// apps/web/types/canva-types.ts
export type ActiveElement = {
  name: string;
  value: string;
  Icon: LucideIcon | null;
} | null;
// ...
export type ShapesMenuProps = {
  item: {
    name: string;
    Icon: LucideIcon;
    value: Array<ActiveElement>;
  };
  activeElement: any;
  handleActiveElement: any;
  handleImageUpload: any;
  imageInputRef: any;
};